Track imagery-search's map view in URL for sharing
The optimal event to listen for changes is 'moveend' but the Mapbox
Synch plugin seems to continuously trigger it. So if the code to update
the URL is also bound to 'moveend' then there is significant excess code
execution -- browser slow down and URL constantly updates. Fortunately
falling back to using a combination of 'dragend' and 'zoomend' seems
sufficient.

Fixes #87

Add throttle to onMapChange()

When using `moveend` to trigger URL updates then you suffer from
continuous triggering from the map sync plugin. So using a bit of
throttling on the callback prevents jitter. However there is now an
artifact where the mouse cursor flashes between grab and cursor.
Nevertheless this currently seems like the most practical and effective
compromise.